learn Authentication & Authorization

+++ PracticeProject-03/TodoApp/router/todos.py
@@ -1,12 +1,11 @@
 from fastapi import APIRouter, Depends, HTTPException, Path
 from typing import Annotated
-
 from pydantic import BaseModel, Field
 from sqlalchemy.orm import Session
 from starlette import status
-
 from models import Todos
 from database import sessionLocal
+from .auth import get_current_user
 
 router = APIRouter()
 
@@ -19,6 +18,7 @@ def get_db():
         db.close()
 
 db_dependency = Annotated[Session, Depends(get_db)]
+user_dependency = Annotated[dict, Depends(get_current_user)]
 
 class TodoRequest(BaseModel):
     title: str = Field(min_length = 3)
@@ -27,25 +27,35 @@ class TodoRequest(BaseModel):
     completed: bool
 
 @router.get("/", status_code=status.HTTP_200_OK)
-async def read_all(db : db_dependency):
-    return db.query(Todos).all()
+async def read_all(user : user_dependency, db : db_dependency):
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Authentication failed")
+    return db.query(Todos).filter(Todos.owner_id == user.get('id')).all()
 
 @router.get("/todo/{todo_id}" , status_code=status.HTTP_200_OK)
-async def read_todo(db : db_dependency, todo_id : int = Path(gt = 0)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+async def read_todo(user : user_dependency,db : db_dependency, todo_id : int = Path(gt = 0)):
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Authentication failed")
+
+    todo_model = db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get('id')).first()
     if todo_model is not None:
         return todo_model
     raise HTTPException(status_code=404, detail="todo not found")
 
 @router.post("/todo", status_code=status.HTTP_201_CREATED)
-async def create_todo(db : db_dependency, todo_request :TodoRequest):
-    todo_model = Todos(**todo_request.model_dump())
+async def create_todo(user : user_dependency,db : db_dependency, todo_request :TodoRequest):
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Authentication failed")
+    todo_model = Todos(**todo_request.model_dump() , owner_id = user.get('id'))
     db.add(todo_model)
     db.commit()
 
 @router.put("/todo/{todo_id}" , status_code=status.HTTP_204_NO_CONTENT)
-async def updated_todo(db : db_dependency, todo_request :TodoRequest, todo_id : int = Path(gt = 0) ):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+async def updated_todo(user: user_dependency, db : db_dependency, todo_request :TodoRequest, todo_id : int = Path(gt = 0) ):
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Authentication failed")
+
+    todo_model = db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get('id')).first()
     if todo_model is None:
         raise HTTPException(status_code=404, detail='todo not found')
     todo_model.title = todo_request.title
@@ -56,8 +66,11 @@ async def updated_todo(db : db_dependency, todo_request :TodoRequest, todo_id : 
     db.commit()
 
 @router.delete("/todo/{todo_id}" , status_code=status.HTTP_204_NO_CONTENT)
-async def delete_todo(db : db_dependency, todo_id : int = Path(gt = 0)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+async def delete_todo(user : user_dependency, db: db_dependency, todo_id : int = Path(gt = 0)):
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Authentication failed")
+
+    todo_model = db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get('id')).first()
     if todo_model is None:
         raise HTTPException(status_code=404 , detail= 'todo not found')
     db.delete(todo_model)
